requirements.txt
# server-build
flask
flask-cors
flask_sqlalchemy
flask_login
# server-tools
gunicorn
requests
Werkzeug
# parser-tools
markdown
PyYAML
python-dotenv
# application-tools
imageio
imageio-ffmpeg
mpmath
bs4
matplotlib
numba
numpy
scipy
wikipedia